add Schedule Hítory

+++ app/src/main/java/com/longhb/do4life/model/retrofit/res/ScheduleHistory.java
@@ -28,7 +28,7 @@ public class ScheduleHistory {
     public String room;
     @SerializedName("time")
     @Expose
-    public Integer time;
+    public long time;
     @SerializedName("result")
     @Expose
     public String result;
